+++ src/main/java/tech/gruppone/stalker/server/model/db/UserDao.java
@@ -14,7 +14,7 @@ import org.springframework.data.relational.core.mapping.Table;
 public class UserDao {
   @Id
   @Column("id")
-  Long id;
+  long id;
 
   @NonNull
   @Column("email")
